+++ app/src/main/java/com/example/rocketapp/model/rocket/info/RocketMass.kt
@@ -1,6 +1,6 @@
 package com.example.rocketapp.model.rocket.info
 
 data class RocketMass(
-    val kg: Int,
-    val lb: Int
+    val kg: Double,
+    val lb: Double
 )
